requirements.txt
absl-py==0.7.1
atari-py==0.0.18
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1
six==1.12.0
Werkzeug==0.15.2
